// types/Generals/sectionHeaderTypes.ts
/** SectionHeader component props interface from path : @/components/Layout/SectionHeader/SectionHeader.tsx */
export interface SectionHeaderProps {
  title1?: string;
  title2?: string;
  subtitle?: string;
  className?: string;
  textAlign?: 'left' | 'center' | 'right';
  titleColor?: string;
  titleLayout?: 'row' | 'column';
}